requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
sentence-transformers==2.3.1
transformers==4.36.2
openai==1.10.0
pandas==2.1.4
numpy==1.24.3
pypdf==3.17.4
python-docx==1.1.0
openpyxl==3.1.2
python-dotenv==1.0.0
pydantic==2.5.3
pyyaml==6.0.1
plotly==5.18.0
matplotlib==3.8.2
requests==2.31.0
tqdm==4.66.1
